debugging ingestion.py/ test_hana.py for testing
