+++ datahub_sap_hana/hana_recipe.yaml
@@ -5,14 +5,32 @@ source:
     password: Localdev1
     database: HXE
     host_port: localhost:39041
+    include_view_lineage: true
     schema_pattern:
-      deny: 
-        - .*SYS*
-    include_view_lineage: false
+      allow:
+        - "HOTEL"
+      ignoreCase: true
+    profile_pattern:
+      allow:
+        - "HOTEL"
+    profiling:
+      enabled: True
+      include_field_null_count: true
+      include_field_min_value: true
+      include_field_max_value: true
+      include_field_mean_value: true
+      include_field_median_value: true
+      include_field_stddev_value: true
+      include_field_quantiles: true
+      include_field_distinct_value_frequencies: true
+      include_field_histogram: true
+      include_field_sample_values: true
+    domain:
+      "urn:li:domain:sales":
+        allow:
+          - "HOTEL"
 sink:
   type: datahub-rest
   config:
-    server: "http://datahub-gms:8080"
-    
-
-    #filename: ./integration_tests/hana_golden_file.json
+    server: $DATAHUB_GMS_URL
+    token: $DATAHUB_GMS_TOKEN
