--- hana_recipe.yaml
@@ -1,37 +1,0 @@
-source:
-  type: datahub_sap_hana.ingestion.HanaSource
-  config:
-    username: HOTEL
-    password: Localdev1
-    database: HXE
-    host_port: localhost:39041
-    include_view_lineage: true
-    include_column_lineage: true 
-    schema_pattern:
-      allow:
-        - "HOTEL"
-      ignoreCase: true
-    profile_pattern:
-      allow:
-        - "HOTEL"
-    profiling:
-      enabled: True
-      include_field_null_count: true
-      include_field_min_value: true
-      include_field_max_value: true
-      include_field_mean_value: true
-      include_field_median_value: true
-      include_field_stddev_value: true
-      include_field_quantiles: true
-      include_field_distinct_value_frequencies: true
-      include_field_histogram: true
-      include_field_sample_values: true
-    domain:
-      "urn:li:domain:sales":
-        allow:
-          - "HOTEL"
-sink:
-  type: datahub-rest
-  config:
-    server: $DATAHUB_GMS_URL
-    token: $DATAHUB_GMS_TOKEN
