+++ tests/integration/data/hana_to_file_lineage_disabled.yml
@@ -1,8 +1,7 @@
 
 # This is copied and slightly modified from the acryl-datahub project
-# Copyright 2015 LinkedIn Corp. All rights reserved.
 # The original source code can be found at https://github.com/acryldata/datahub/blob/master/metadata-ingestion/tests/integration/hana/hana_to_file.yml
-# And the original Apache 2 license is available at https://github.com/acryldata/datahub/blob/master/LICENSE
+# And the license is available at XXXX
 
 run_id: hana-test
 source:
